requirements.txt
discord.py==1.7.3
python-dotenv==0.19.2
pyracing==0.1.1